New Array and Finished Functions
Prompts did not work without a new array named choice. All decision made through the prompts will be compiled into a choice which the generatePassword function can use to make the actual password. There was no connection between the prompts and generating the password before. The generatePassword function can now use the data from the prompts to make a password.